FirestoreDataManager refactoring
 - to separate files
 - Channel and Message extension for fields string representations
